+++ docs/public/domnit.js
@@ -8260,6 +8260,7 @@ Domnit = (function() {
     attributes are considered inherited by the W3C (not guarenteed to be the behavior of every browser).
     Default: `font-family`, `font-size`, `font-style`, `font-variant`, `font-weight`, `line-height`, `letter-spacing`,
     `visibility`.
+  @option opt [int] concurrency Define how many serializations can occur concurrently.  Defaults to `Infinity`.
    */
   function Domnit(opt) {
     this.opt = opt != null ? opt : {};
@@ -8280,7 +8281,8 @@ Domnit = (function() {
       filterHidden: false,
       filterDisplayNone: false,
       inheritStyle: true,
-      inheritSilent: ["font-family", "font-size", "font-style", "font-variant", "font-weight", "line-height", "letter-spacing", "visibility"]
+      inheritSilent: ["font-family", "font-size", "font-style", "font-variant", "font-weight", "line-height", "letter-spacing", "visibility"],
+      concurrency: 2
     });
   }
 
@@ -8346,7 +8348,7 @@ Domnit = (function() {
    */
 
   Domnit.prototype.serialize = function(el, root) {
-    var Serializer, child, children, customSerialize, i, len, ref, ref1;
+    var Serializer, customSerialize, ref;
     if (root == null) {
       root = true;
     }
@@ -8355,13 +8357,13 @@ Domnit = (function() {
     }
     customSerialize = (el.tagName.toLowerCase()) + "Serializer";
     Serializer = (ref = this[customSerialize]) != null ? ref : this.elementSerializer;
-    children = [];
-    ref1 = el.children;
-    for (i = 0, len = ref1.length; i < len; i++) {
-      child = ref1[i];
-      children.push(this.serialize(child, false));
-    }
-    return Promise$1.all(children).then((function(_this) {
+    return Promise$1.map(el.children, ((function(_this) {
+      return function(child) {
+        return _this.serialize(child, false);
+      };
+    })(this)), {
+      concurrency: this.opt.concurrency
+    }).then((function(_this) {
       return function(children) {
         var element;
         element = new Serializer(el, children, _this.opt, root);