+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## Unreleased
 
 ### Added
-- CodeMirror demonstration
+- CodeMirror [demonstration][demo:codemirror]
 - [Domnit][Domnit#constructor] takes `concurrency` option, limiting how much Domnit blocks other processes.
 - [Domnit][Domnit#constructor] takes `nonBlocking` option, requiring Domnit to give up process control.
+
+### Modified
+- **[edge-case]** `nonBlocking` defaults to `true`, allowing other actions to take place on the page.  If you depended
+  on other actions waiting until Domnit finished, set `nonBlocking` to `false`.
 
 ## [0.3.0] - 2016-11-04
 
@@ -39,3 +43,4 @@
 [ElementSerializer#update]: https://rweda.github.io/gaddomnit/#https://rweda.github.io/gaddomnit/class/ElementSerializer.html#update-dynamic
 [ElementSerializer#interweaveText]: https://rweda.github.io/gaddomnit/#https://rweda.github.io/gaddomnit/class/ElementSerializer.html#interweaveText-dynamic
 [DefaultStyle.getNoCache]: https://rweda.github.io/gaddomnit/#https://rweda.github.io/gaddomnit/class/DefaultStyle.html#getNoCache-dynamic
+[demo:codemirror]: https://rweda.github.io/gaddomnit/demo/codemirror.html